+++ gemalaya/qml/GemSpaceNextAction.qml
@@ -4,7 +4,7 @@ import QtQuick.Controls 2.14
 Action {
   property Item stackLayout
 
-  shortcut: Conf.shortcuts.stackNext
+  shortcut: Conf.stackShortcuts.stackNext
 
   onTriggered: {
     if (stackLayout.currentIndex < (stackLayout.count - 1))